Update dependencies and refactor import paths for controller-util package
