+++ pkg/controller/mysqlbackup/internal/syncer/deletionjob_test.go
@@ -23,7 +23,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	syncerpkg "github.com/presslabs/controller-util/syncer"
+	syncerpkg "github.com/presslabs/controller-util/pkg/syncer"
 	batch "k8s.io/api/batch/v1"
 	core "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
