+++ pkg/controller/mysqlbackup/internal/syncer/syncer_suite_test.go
@@ -28,7 +28,6 @@ import (
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 
 	api "github.com/bitpoke/mysql-operator/pkg/apis/mysql/v1alpha1"
 )
@@ -39,7 +38,7 @@ var c client.Client
 
 func TestSyncers(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecsWithDefaultAndCustomReporters(t, "Backup syncers suite", []Reporter{printer.NewlineReporter{}})
+	RunSpecs(t, "Backup syncers suite")
 }
 
 var _ = BeforeSuite(func() {
