+++ pkg/controller/mysqlbackup/mysqlbackup_controller_suite_test.go
@@ -26,7 +26,6 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/bitpoke/mysql-operator/pkg/apis"
@@ -38,7 +37,7 @@ var t *envtest.Environment
 
 func TestMysqlBackupController(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecsWithDefaultAndCustomReporters(t, "MysqlBackup Controller Suite", []Reporter{printer.NewlineReporter{}})
+	RunSpecs(t, "MysqlBackup Controller Suite")
 }
 
 var _ = BeforeSuite(func() {
