+++ pkg/controller/mysqlcluster/internal/syncer/config_map.go
@@ -29,7 +29,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/presslabs/controller-util/syncer"
+	"github.com/presslabs/controller-util/pkg/syncer"
 
 	"github.com/bitpoke/mysql-operator/pkg/internal/mysqlcluster"
 )
