+++ pkg/controller/mysqlcluster/internal/syncer/healthy_service.go
@@ -22,7 +22,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/presslabs/controller-util/syncer"
+	"github.com/presslabs/controller-util/pkg/syncer"
 
 	"github.com/bitpoke/mysql-operator/pkg/internal/mysqlcluster"
 )
