+++ pkg/controller/mysqlcluster/internal/syncer/mysqlcluster_suite_test.go
@@ -27,7 +27,6 @@ import (
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 
 	api "github.com/bitpoke/mysql-operator/pkg/apis/mysql/v1alpha1"
 )
@@ -38,7 +37,7 @@ var c client.Client
 
 func TestSyncers(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecsWithDefaultAndCustomReporters(t, "Syncers suit", []Reporter{printer.NewlineReporter{}})
+	RunSpecs(t, "Syncers suit")
 }
 
 var _ = BeforeSuite(func() {
