+++ pkg/controller/mysqlcluster/internal/syncer/secret.go
@@ -19,7 +19,7 @@ package mysqlcluster
 import (
 	"fmt"
 
-	"github.com/presslabs/controller-util/syncer"
+	"github.com/presslabs/controller-util/pkg/syncer"
 	core "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
