+++ pkg/controller/mysqlcluster/mysqlcluster_controller.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"reflect"
 
-	"github.com/presslabs/controller-util/syncer"
+	"github.com/presslabs/controller-util/pkg/syncer"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	policyv1 "k8s.io/api/policy/v1"
