+++ pkg/controller/mysqldatabase/db_controller_suite_test.go
@@ -25,7 +25,6 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/bitpoke/mysql-operator/pkg/apis"
@@ -37,7 +36,7 @@ var t *envtest.Environment
 
 func TestMySQLDatabase(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecsWithDefaultAndCustomReporters(t, "MySQL Database Suite", []Reporter{printer.NewlineReporter{}})
+	RunSpecs(t, "MySQL Database Suite")
 }
 
 var _ = BeforeSuite(func() {
