+++ pkg/controller/mysqluser/mysqluser_controller.go
@@ -24,8 +24,8 @@ import (
 	"time"
 
 	"github.com/go-test/deep"
-	logf "github.com/presslabs/controller-util/log"
-	"github.com/presslabs/controller-util/meta"
+	logf "github.com/presslabs/controller-util/pkg/log"
+	"github.com/presslabs/controller-util/pkg/meta"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
