+++ pkg/controller/mysqluser/mysqluser_controller_suite_test.go
@@ -25,7 +25,6 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/bitpoke/mysql-operator/pkg/apis"
@@ -37,7 +36,7 @@ var t *envtest.Environment
 
 func TestMySQLUserController(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecsWithDefaultAndCustomReporters(t, "MysqlUser Controller Suite", []Reporter{printer.NewlineReporter{}})
+	RunSpecs(t, "MysqlUser Controller Suite")
 }
 
 var _ = BeforeSuite(func() {
