+++ pkg/controller/orchestrator/orchestrator_controller.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/go-test/deep"
-	"github.com/presslabs/controller-util/syncer"
+	"github.com/presslabs/controller-util/pkg/syncer"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
