+++ pkg/controller/orchestrator/orchestrator_controller_suite_test.go
@@ -27,7 +27,6 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/bitpoke/mysql-operator/pkg/apis"
@@ -39,7 +38,7 @@ var t *envtest.Environment
 
 func TestMysqlClusterController(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecsWithDefaultAndCustomReporters(t, "Orchestrator Controller Suite", []Reporter{printer.NewlineReporter{}})
+	RunSpecs(t, "Orchestrator Controller Suite")
 }
 
 var _ = BeforeSuite(func() {
