+++ pkg/orc-helper/helper.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/presslabs/controller-util/rand"
+	"github.com/presslabs/controller-util/pkg/rand"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
